requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
filelock>=3.13.0
loguru>=0.7.0
